MSAPI-96 | Rename Constants
